Album art downloading first pass
Unfortunetly I am visiting relatives for the next two weeks starting
tomorrow so I will be on another machine thus I need to commit this.
It is still buggy and I think I am just going to have to rethink the
write through flagging as if the album art changes it is not flagged. It
will write to disk however the album art on active music player change.