Added progress support & cover fixes
Added progress support to the plugin, I may make it that webpages are
able to send it if they have it precalculated but I am not sure yet.
Also I am thinking about adding ms support even though not every
platform will support it,.

Fixed a few edge cases for cover not downloading such as closing a
connection. Also added a likely unneeded check to force download if
empty.