Remove spotify exceptions, new SetPosition command
SetPosition now internally sends both the seconds and percent (0-1) of
where the song needs to be set to. It is the up to the script to decide
which is best to use for it. I also remove the exceptions to only do
that if you are a spotify premium user since I no longer user their API
for that.